# backend-go/api-spec.yml
openapi: 3.0.3
info:
  title: Email Helper API
  description: REST API for Email Helper application with AI-powered email classification and task management
  version: 1.0.0
  contact:
    name: Email Helper
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/ai/classify:
    post:
      summary: Classify email content
      tags: [AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subject
                - sender
                - content
              properties:
                subject:
                  type: string
                  example: "Team meeting tomorrow"
                sender:
                  type: string
                  example: "manager@company.com"
                content:
                  type: string
                  example: "Please join us for the team meeting tomorrow at 2 PM"
                context:
                  type: string
                  example: ""
      responses:
        '200':
          description: Classification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailClassificationResponse'
        '500':
          description: AI processing failed
  /api/ai/action-items:
    post:
      summary: Extract action items from email
      tags: [AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email_content
              properties:
                email_content:
                  type: string
                context:
                  type: string
      responses:
        '200':
          description: Action items extracted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionItemResponse'
  /api/ai/summarize:
    post:
      summary: Generate email summary
      tags: [AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email_content
              properties:
                email_content:
                  type: string
                summary_type:
                  type: string
                  enum: [brief, detailed]
                  default: brief
      responses:
        '200':
          description: Summary generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummaryResponse'
...
components:
  schemas:
...
    EmailClassificationResponse:
      type: object
      properties:
        category:
          type: string
          enum:
            - required_personal_action
            - team_action
            - optional_event
            - fyi
            - newsletter
            - job_listing
            - spam_to_delete
            - work_relevant
        confidence:
          type: number
          minimum: 0
          maximum: 1
        reasoning:
          type: string
        alternative_categories:
          type: array
          items:
            type: string
        processing_time:
          type: number
        one_line_summary:
          type: string
    ActionItemResponse:
      type: object
      properties:
        action_items:
          type: array
          items:
            type: string
        urgency:
          type: string
        deadline:
          type: string
        confidence:
          type: number
        due_date:
          type: string
        action_required:
          type: boolean
        explanation:
          type: string
        relevance:
          type: string
        links:
          type: array
          items:
            type: string
    SummaryResponse:
      type: object
      properties:
        summary:
          type: string
        key_points:
          type: array
          items:
            type: string
        confidence:
          type: number
        processing_time:
          type: number
...